Use Path/PathBuf instead of String in file_formatter
This avoids a panic if the underlying paths are not representable in
utf8. As a bonus, we can use `Path::metadata` to check if a path is a
directory or file with one OS interaction instead of two.

When printing error messages, we have to convert to utf8 which means
potentially introducing some replacement characters for non-utf8 paths.
This is better than panicking though.

One other slight change in behaviour that comes from using the path API
more is what we treat as a matching file extension. In particular '.pas'
is no longer considered a formattable path.