Only write to files when the contents need changing
One of the main use cases for this tool is incremental formatting on
large repositories. In such cases, writing to every file on every format
is quite wasteful because it requires the version control system to
re-index all files. To save this time in VCS indexing, we can spend a
little time checking whether a file needs changing before writing to it.

This string comparison is comparatively very fast and significantly
improves performance in the incremental case (15-40% faster overall in
our benchmarks on my machine). When files need to be changed it should
be slower overall, but the impact was not measurable on my machine.