Add `string` keyword
Initially we thought that we wouldn't treat `string` as a keyword
because we wanted to have it formatted like an identifier. This was a
bad idea because it's much more important to have the core tool classify
tokens correctly, and it's always possible for a formatter to decide to
make a special case for `string`.