Tweak warning for `Unknown` token
The leading whitespace was being printed before the unexpected
character, and it could print some of the following source code which
won't really ever be useful since it's just the one character that is
unexpected. If we want to help the user locate the unknown token we
should just given them the file name and position.