Switch to `windows-sys`
It's a lightweight version of the `windows` crate that only provides
the function declarations for calling external functions, whereas the
`windows` crate provides some convenience helpers, traits, and wrappers.
Like the `windows` crate, it's provided by Microsoft themselves.

For us it's a drop-in replacement that should only reduce compile time
(especially because we were already dependening on `windows-sys`
transitively).