Remove need for `Serialize` bound in `get_config_object`
It was only used for logging the final configuration object, but the
caller can do this themselves (and choose the representation).

This also changes the logging of the configuration object from the toml
file representation to a simpler struct-level debug printing. Since it's
just a debugging feature, it's fine to use the simpler printing,
especially when it cuts back on the generated code and cuts one use of a
dependency. This also fixes the annoying issue with '\r\n' being printed
with inconsistent escaping.