Panic on whitespace count overflow
The current approach of truncating the whitespace count is unsound. It's
possible for the offending token to be following a single-line comment
in which case the trailing newline for the comment will be truncated,
which could cause code to become commented out.

Since it's already quite unrealistic for this whitespace overflow to
occur (2^31 bytes - 4 GB - of whitespace is required), it seems fine to
just crash the program.