Fix web demo build
For some reason, after the clap feature change in #129 the build of the
web demo is broken. What's even more strange is that it only fails at
build time (not at type-check time). Maybe it's because of some build
script that no longer compiles?

In any case, upgrading both `clap` and `windows-sys` fixes the problem.