Fix optimisation bug around best penalties
When traversing nodes in the solution space, there is a concept of
"best penalties". This is maintained at the points where there is a
`MustBreak` requirement.

Previously, the penalty that was being inserted into the list was the
penalty of the solution before adding the `MustBreak` decision's
penalty. Additionally, when the penalty is updated, there can be
multiple solutions with differing penalties. All of these need to be
accounted for in the "best penalty".