Stop deleting voided logical lines
By removing the "voided" logical lines, it is necessary to rewrite all
the paren line indices of each line. Rather than doing this, it was
thought to be okay to just allow formatters to have to handle these
empty lines.

They can either handle empty lines explicitly, or early exit when
dealing with lines with type "Voided".