Add safety check for line breaks after single-line comments
As discovered recently (#181), it's possible for subtle bugs to slip in
that result in a missing line break after a single-line comment.

While we strive to never let this happen, it's a good idea to make
*sure* that it doesn't happen. With this change, the missing line break
will be detected and inserted. The formatting won't be great after a
line break is suddenly inserted, but that's better than broken.