Exclude test code from coverage
Including tests is artificially on the whole increasing our reported
code coverage, and in some files it creates lines that can only be
covered when tests fail.

This change applies the suggestion given in
https://github.com/taiki-e/cargo-llvm-cov?tab=readme-ov-file#exclude-code-from-coverage
to exclude coverage in a specific module.

An alternative solution to this is to move all test code into a
directory named tests, or a file called `tests.rs`. I think this is the
lesser solution, because I like having the test code inline, usually.

Surprisingly, even though the `#[coverage]` attribute that this
(conditionally) uses is unstable, using the nightly toolchain doesn't
appear to be necessary to get `cargo-llvm-cov` to exclude the test code
with this change.